rename gateway_route.py to http_route.py
